chore: rename module and move type (#2683)
## What

This renames `src/api/index.rs` to `src/api/builder_fns.rs` and then
moves the `struct FieldName` type and its associated trait impls and
functions to `src/api/mod.rs`.
