+++ pg_search/src/postgres/customscan/builders/custom_path.rs
@@ -15,8 +15,8 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-use crate::api::index::FieldName;
 use crate::api::Cardinality;
+use crate::api::FieldName;
 use crate::api::HashSet;
 use crate::index::fast_fields_helper::WhichFastField;
 use crate::postgres::customscan::CustomScan;
